+++ modules/features/profile/src/main/java/au/com/shiftyjelly/pocketcasts/profile/cloud/CloudFilesViewModel.kt
@@ -4,36 +4,21 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.toLiveData
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
-import au.com.shiftyjelly.pocketcasts.analytics.EpisodeAnalytics
-import au.com.shiftyjelly.pocketcasts.analytics.SourceView
-import au.com.shiftyjelly.pocketcasts.models.entity.BaseEpisode
-import au.com.shiftyjelly.pocketcasts.models.entity.UserEpisode
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.repositories.file.CloudFilesManager
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
-import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.UserEpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.user.UserManager
-import au.com.shiftyjelly.pocketcasts.views.helper.CloudDeleteHelper
-import au.com.shiftyjelly.pocketcasts.views.helper.DeleteState
-import au.com.shiftyjelly.pocketcasts.views.helper.EpisodeItemTouchHelper.SwipeAction
-import au.com.shiftyjelly.pocketcasts.views.helper.EpisodeItemTouchHelper.SwipeSource
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class CloudFilesViewModel @Inject constructor(
     private val userEpisodeManager: UserEpisodeManager,
     private val playbackManager: PlaybackManager,
-    private val episodeManager: EpisodeManager,
     private val settings: Settings,
     userManager: UserManager,
     private val analyticsTracker: AnalyticsTrackerWrapper,
-    private val episodeAnalytics: EpisodeAnalytics,
     private val cloudFilesManager: CloudFilesManager,
 ) : ViewModel() {
 
@@ -57,46 +42,6 @@ class CloudFilesViewModel @Inject constructor(
         userEpisodeManager.syncFilesInBackground(playbackManager)
     }
 
-    @OptIn(DelicateCoroutinesApi::class)
-    fun episodeSwipeUpNext(episode: BaseEpisode) {
-        GlobalScope.launch(Dispatchers.Default) {
-            if (playbackManager.upNextQueue.contains(episode.uuid)) {
-                playbackManager.removeEpisode(episodeToRemove = episode, source = SourceView.FILES)
-                trackSwipeAction(SwipeAction.UP_NEXT_REMOVE)
-            } else {
-                playbackManager.playNext(episode = episode, source = SourceView.FILES)
-                trackSwipeAction(SwipeAction.UP_NEXT_ADD_TOP)
-            }
-        }
-    }
-
-    @OptIn(DelicateCoroutinesApi::class)
-    fun episodeSwipeUpLast(episode: BaseEpisode) {
-        GlobalScope.launch(Dispatchers.Default) {
-            if (playbackManager.upNextQueue.contains(episode.uuid)) {
-                playbackManager.removeEpisode(episodeToRemove = episode, source = SourceView.FILES)
-                trackSwipeAction(SwipeAction.UP_NEXT_REMOVE)
-            } else {
-                playbackManager.playLast(episode = episode, source = SourceView.FILES)
-                trackSwipeAction(SwipeAction.UP_NEXT_ADD_BOTTOM)
-            }
-        }
-    }
-
-    fun getDeleteStateOnSwipeDelete(episode: UserEpisode): DeleteState {
-        trackSwipeAction(SwipeAction.DELETE)
-        return CloudDeleteHelper.getDeleteState(episode)
-    }
-
-    fun deleteEpisode(episode: UserEpisode, deleteState: DeleteState) {
-        CloudDeleteHelper.deleteEpisode(episode, deleteState, playbackManager, episodeManager, userEpisodeManager)
-        episodeAnalytics.trackEvent(
-            event = if (deleteState == DeleteState.Cloud && !episode.isDownloaded) AnalyticsEvent.EPISODE_DELETED_FROM_CLOUD else AnalyticsEvent.EPISODE_DOWNLOAD_DELETED,
-            source = SourceView.FILES,
-            uuid = episode.uuid,
-        )
-    }
-
     fun changeSort(sortOrder: Settings.CloudSortOrder) {
         settings.setCloudSortOrder(sortOrder)
         cloudFilesManager.sortOrderRelay.accept(sortOrder)
@@ -104,16 +49,6 @@ class CloudFilesViewModel @Inject constructor(
 
     fun getSortOrder(): Settings.CloudSortOrder {
         return cloudFilesManager.sortOrderRelay.value ?: settings.getCloudSortOrder()
-    }
-
-    private fun trackSwipeAction(swipeAction: SwipeAction) {
-        analyticsTracker.track(
-            AnalyticsEvent.EPISODE_SWIPE_ACTION_PERFORMED,
-            mapOf(
-                ACTION_KEY to swipeAction.analyticsValue,
-                SOURCE_KEY to SwipeSource.FILES.analyticsValue
-            )
-        )
     }
 
     companion object {