+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 -----
 *   New Feature:
     *   Added 3 episodes on the Filter AutoDownload
-        ([#1169])(https://github.com/Automattic/pocket-casts-android/pull/1169)
+        ([#1169](https://github.com/Automattic/pocket-casts-android/pull/1169))
     *   Added capability to deselect all/below and above on the multiselect feature
         ([#1172](https://github.com/Automattic/pocket-casts-android/pull/1172))
+    *   Added share option to episode swipe and multiselect menus
+        ([#1190](https://github.com/Automattic/pocket-casts-android/pull/1190)),
+        ([#1191](https://github.com/Automattic/pocket-casts-android/pull/1191))
 
 7.43
 -----